Create STARTER
In this project, I use the main.py file as the starting point of the TRESSÉ Hair Salon Appointment Management System. This file is responsible for launching the application, creating the main graphical user interface using Tkinter, and connecting the user interface to the system’s backend logic. All program execution begins in this file.